test: updates add crud coverage for tracing
